Move GeodeticDatum to geodesy module
