+++ src/navtools/src/navtools/conversions/coordinates.py
@@ -12,7 +12,7 @@ from typing import NamedTuple
 
 import numpy as np
 
-from navtools.constants import GeodeticDatum
+from navtools.geodesy import GeodeticDatum
 
 
 class ECI(NamedTuple):
@@ -174,7 +174,9 @@ def ecef2geodetic(
     lon = np.arctan2(y, x)
     alt = (beta - datum.r0 * T) * np.cos(lat) + (
         z - np.sign(z) * datum.r0 * np.sqrt(1 - datum.eccentricity**2)
-    ) * np.sin(lat)  # Eq. C.38
+    ) * np.sin(
+        lat
+    )  # Eq. C.38
 
     return GEODETIC(lat=lat, lon=lon, alt=alt)
